Version Beta 1.69
vb1.69

Changes:
- Added MuzzleFlash to Settings -> Options
- Added NoFog to Settings -> Options
- Added MagazineCasings to Settings -> Options

- I temporarily set the value of "Auto-Update" to false while I am still working on the patcher overhaul.
If you still want to have automatic updates, you can enable the option in the launcher settings. You will also see in the top right corner of the launcher if an update is available. Just activate "auto-update", restart the launcher, and if necessary, deactivate the function again.

Note: If the "NoFog" checkbox is enabled, it means that fog is enabled in the game. If it is disabled, then there is no fog in the game. I will change this in the next update to avoid confusion.

Haven't tested all the changes in case there are errors. Please let me know immediately so I can fix it.